It would seem to be working fine, but...
But in reality the logic is incorrect.
I thought it would be neat to base the calculation on the "physics ticks",
or scoring_ticks,  that is, a counter that's incremented every time
the Gmotor engine calls our scoring routine.

That doesn't work unfortunately, and it's clear when you stop the
car on track. At that point, the current lap ticks go on and on,
but there's no corresponding ticks for the best lap, so things tend
to explode.

Correct logic must be based on track position alone (mLapDist),
not on ticks or time.

Included a few log files to help understand the inner working.