Color of delta time is now based on derivative
This is the way iRacing works, I'm told.
If you're gaining time (even if you are slower, i.e. in the plus zone)
the time will be green.
If you're losing time compared to your best lap (even if you're faster,
i.e. in the minus zone), the time will be red.

I have to say I don't like how this *looks* very much, but
it does convey the message.